backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
tavily-python>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
duckduckgo-search>=6.0.0
